src/web/portfolio/project/api.py
from rest_framework import viewsets
from .models import Technology, Project
from .serializers import TechnologySerializer, ProjectSerializer


class TechnologyViewSet(viewsets.ModelViewSet):
    queryset = Technology.objects.all()
    serializer_class = TechnologySerializer


class ProjectViewSet(viewsets.ModelViewSet):
    queryset = Project.objects.all()
    serializer_class = ProjectSerializer
